Add custom ContextDatabase implementation
    In TranslationContext::translate(), the call to Sleigh::initialize()
for resetting the translation engine leads to an unhandled exception
in ContextInternal::registerVariable() under certain circumstances.

There is no way to reset a ContextInternal database short of creating a
new one for every call to translate().

Add a ContextInternal subclass that ignores registering variables after
the first initialization on the sleigh engine, and supports resetting
the values of context variables to their default values.